fix: Fix alert detection, reorganize files
