+++ listeners/bot_insult.py
@@ -1,0 +1,38 @@
+import discord
+from discord.ext import commands
+from discord import AllowedMentions
+import logging
+import random
+from utils.config_loader import SETTINGS_DATA, MISC_DATA
+
+logger = logging.getLogger('endurabot.' + __name__)
+
+class bot_insult(commands.Cog):
+    def __init__(self, bot):
+        self.bot = bot
+        self.settings_data = SETTINGS_DATA
+        self.misc_data = MISC_DATA        
+    
+        self.default_allowed_mentions = AllowedMentions(
+                everyone=False,
+                users=True, 
+                roles=True      
+            )
+
+    @commands.Cog.listener()
+    async def on_message(self, message: discord.Message):
+
+        chance = SETTINGS_DATA["bot_insult_chance"]
+
+        if self.bot.user.mentioned_in(message):
+
+            value = random.random()
+
+            if value <= chance:
+                random_insult = random.choice(MISC_DATA["bot_insults"])
+                await message.channel.send(f"{message.author.mention}, {random_insult}", allowed_mentions=self.default_allowed_mentions)
+            else:
+                return 
+
+async def setup(bot):
+    await bot.add_cog(bot_insult(bot))
